Use proper type for approxDimension commit_hash:823edf204983811bb5515a4def6185ddffd21578

+++ catboost/private/libs/algo_helpers/approx_updater_helpers.h
@@ -236,6 +236,6 @@ void InitApproxFromBaseline(
 void InitApproxes(
     int size,
     const TMaybe<TVector<double>>& startingApprox,
-    double approxDimension,
+    size_t approxDimension,
     bool storeExpApproxes,
     TVector<TVector<double>>* approx);
